+++ examples/app/mcorder.js
@@ -1,4 +1,14 @@
-angular.module('McOrderApp', ['ui.directives', 'StarcounterLib']);
+angular.module('McOrderApp', ['ui.directives', 'StarcounterLib'])
+  .config(['$routeProvider', function($routeProvider) {
+    $routeProvider.
+    when('/grid', {
+      templateUrl: '../examples/app/mcorder_grid.html', 
+      controller: MyCtrl
+    }).
+    otherwise({
+      redirectTo: '/grid'
+    });
+  }]);
 
 function MyCtrl($scope, $http, $location) {
   $scope.dumpItems = function(){